fix: clear +10 cppcheck ratchet findings from merged NAS decoders (#177)
- decoder.h: cast CHECK_PDU_POINTER_AND_LENGTH_DECODER's LOG length args to
  (int) — call sites pass uint32_t len into a %d format (invalidPrintfArgType_
  sint in eps_quality_of_service, esp_mobile_identity, nas_tracking_area_
  identity, pdn_address, tracking_area_identity_list)
- nas_msg.c: drop the redundant buffer==NULL half of the second
  CHECK_PDU_POINTER_AND_LENGTH_DECODER (buffer is already dereferenced after
  the first check) — use CHECK_LENGTH_DECODER for the length leg

+++ src/mmt_mobile/nas/util/decoder.h
@@ -107,7 +107,7 @@ extern int errorCodeDecoder;
 		while (unlikely( length < minimumlength))                              \
         {                                                                      \
                 LOG("(%s:%d) Expecting at least %d bytes, got %d\n",           \
-                      __FILE__, __LINE__, minimumlength, length);              \
+                      __FILE__, __LINE__, (int) minimumlength, (int) length);  \
                 errorCodeDecoder = DECODE_BUFFER_TOO_SHORT;                    \
                 return (DECODE_BUFFER_TOO_SHORT);                              \
         }
